+++ src/libxtp/qmpackages/nwchem.cc
@@ -584,15 +584,15 @@ namespace votca {
             // copying energies to a matrix
             _orbitals->MOEnergies().resize(_levels);
             //_level = 1;
-            for (size_t i = 0; i < _orbitals->MOEnergies().size(); i++) {
+            for (int i = 0; i < _orbitals->MOEnergies().size(); i++) {
                 _orbitals->MOEnergies()[i] = _energies[ i ];
             }
 
 
             // copying orbitals to the matrix
             (_orbitals->MOCoefficients()).resize(_levels, _basis_size);
-            for (size_t i = 0; i < _orbitals->MOCoefficients().rows(); i++) {
-                for (size_t j = 0; j < _orbitals->MOCoefficients().cols(); j++) {
+            for (int i = 0; i < _orbitals->MOCoefficients().rows(); i++) {
+                for (int j = 0; j < _orbitals->MOCoefficients().cols(); j++) {
                     _orbitals->MOCoefficients()(j, i) = _coefficients[i][j];
                     //cout << i << " " << j << endl;
                 }
